Change type of N param from uint64_t to size_t

+++ crypto_scrypt-nosse.c
@@ -232,7 +232,7 @@ smix(uint8_t * B, size_t r, uint64_t N, uint32_t * V, uint32_t * XY)
  */
 int
 libscrypt_scrypt(const uint8_t * passwd, size_t passwdlen,
-    const uint8_t * salt, size_t saltlen, uint64_t N, uint32_t r, uint32_t p,
+    const uint8_t * salt, size_t saltlen, size_t N, uint32_t r, uint32_t p,
     uint8_t * buf, size_t buflen)
 {
 	void * B0, * V0, * XY0;
